pyproject.toml
[project]
name = "DataPipes"
version = "0.1.0"
description = "DataPipes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "einops>=0.8.1",
    "fastplotlib[imgui]>=0.5.1",
    "h5py>=3.14.0",
    "hdf5plugin>=5.1.0",
    "hickle>=5.0.3",
    "icecream>=2.1.8",
    "imageio[ffmpeg]>=2.37.0",
    "imgui-bundle<1.92",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "jupyter-rfb>=0.5.3",
    "matplotlib>=3.10.6",
    "numcodecs>=0.16.3",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "pyglm>=2.8.2",
    "rich>=14.2.0",
    "scikit-image>=0.25.2",
    "setuptools>=80.9.0",
    "simplejpeg>=1.9.0",
    "tqdm>=4.67.1",
    "zarr>=3.1.3",
    "ipykernel>=6.30.1",
    "nvidia-nvimgcodec-cu12>=0.6.1.37",
    "ndindex>=1.10.0",
    "av>=16.0.1",
    "pynvvideocodec>=2.0.3",
]

#[project.urls]
## This is critical for uv/pip to locate non-PyPI wheels
## uv understands --extra-index-url via dependency-url mapping
#"torch" = "https://download.pytorch.org/whl/cu128"
#"torchvision" = "https://download.pytorch.org/whl/cu128"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128" },
]
torchvision = [
    { index = "pytorch-cu128" },
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/datapipes"]

